🐛 fix(review-api-fix): review list 리스폰스 변경 (NAILCASE-315)
- response 디테일 수정

resolves: [NAILCASE-315](https://nailcase.atlassian.net/browse/NAILCASE-315)
