+++ src/main/java/com/nailcase/config/SecurityConfig.java
@@ -28,9 +28,6 @@ import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-import com.fasterxml.jackson.core.SerializableString;
-import com.fasterxml.jackson.core.io.CharacterEscapes;
-import com.fasterxml.jackson.core.io.SerializedString;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nailcase.exception.codes.AuthErrorCode;
 import com.nailcase.exception.codes.ErrorResponse;
@@ -136,26 +133,6 @@ public class SecurityConfig {
 	@Bean
 	public AuditorAware<Long> auditorProvider() {
 		return new AuditorAwareImpl();
-	}
-
-	@Bean
-	public ObjectMapper objectMapper() {
-		ObjectMapper objectMapper = new ObjectMapper();
-		objectMapper.getFactory().setCharacterEscapes(new CharacterEscapes() {
-			@Override
-			public int[] getEscapeCodesForAscii() {
-				return CharacterEscapes.standardAsciiEscapesForJSON();
-			}
-
-			@Override
-			public SerializableString getEscapeSequence(int ch) {
-				if (ch > 127) {
-					return new SerializedString(String.format("\\u%04x", ch));
-				}
-				return null;
-			}
-		});
-		return objectMapper;
 	}
 
 	@Bean
