+++ src/main/resources/application.properties
@@ -46,3 +46,5 @@ validation.rules.point.msg=Point must be an array of two elements with valid lat
 # response encoding
 server.servlet.encoding.charset=UTF-8
 server.servlet.encoding.force=true
+# jackson
+spring.jackson.serialization.WRITE_CHAR_ARRAYS_AS_JSON_ARRAYS=true
